Add files via upload
This project was done for ELEC-E5550 - Statistical Natural Language Processing course. In this project, two types of experiments were carried out. The first experiment was using the tfidf representation with different machine learning models and also checking
their interpretability with LIME. The second experiment was using word2vec with fastText and checking the accuracy with different values of hyperparameters.